pyproject.toml
[project]
name = "aws-agent-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=1.0.3",
    "bedrock-agentcore-starter-toolkit>=0.1.25",
    "mcp>=1.17.0",
    "python-dotenv>=1.1.1",
    "strands-agents>=1.12.0",
    "strands-agents-tools>=0.2.11",
]
